Move serde-related Dictionary tests to the serde_tests module
